requirements.txt
#requirements.txt
streamlit>=1.28.0
PyMuPDF>=1.23.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
google-generativeai>=0.3.0
langchain>=0.0.350
numpy>=1.24.0
torch>=2.0.0
transformers>=4.35.0
pandas>=2.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
opencv-python>=4.8.0